Fix: Correctly handle locale-specific decimal separators
BUG: 507525
BUG: 501116
BUG: 501190

Fixes a bug where locales that use a comma for decimals, like de_DE, would break calculations. The comma wasn't being converted to a period for the math engine, leading to it being misinterpreted and causing weird results. This now handles the conversion before and after the calculation.

This was originally created as
https://invent.kde.org/utilities/kalk/-/merge_requests/118, but was
updated by @devinlin to expedite getting the fix in due to its severity.