Remove page transition animation
I originally introduced it a few years ago as a way to make switching
pages feel more responsive for touch devices. However, upon reflection,
it adds unnecessary friction for users that are switching pages.
Removing it makes the application feel faster, and also removes
unnecessary complexity in pages to accomplish the animation.